sync_mysql_x/requirements.txt
nameko==2.14.1
mysql-connector-python==8.3.0
